Postopedia User Roles working

+++ app/controllers/wikis_controller.rb
@@ -1,6 +1,5 @@
 class WikisController < ApplicationController
-  #before_action :authenticate_user, except: [:index, :show]
-  before_action :authorize_user, except: [:index, :show]
+  after_action :verify_authorized, except: [:index]
 
   def index
     @wikis = Wiki.all
@@ -8,15 +7,18 @@ class WikisController < ApplicationController
 
   def show
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
   end
 
   def new
     @wiki = Wiki.new
+    authorize @wiki
   end
 
   def create
     @wiki = Wiki.new(wiki_params)
     @wiki.user = current_user
+    authorize @wiki
 
     if @wiki.save
       flash[:notice] = "Wiki was saved."
@@ -30,6 +32,7 @@ class WikisController < ApplicationController
   def update
     @wiki = Wiki.find(params[:id])
     @wiki.assign_attributes(wiki_params)
+    authorize @wiki
 
     if @wiki.save
       flash[:notice] = "Wiki was updated."
@@ -42,14 +45,16 @@ class WikisController < ApplicationController
 
   def edit
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
   end
 
   def destroy
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
 
     if @wiki.destroy
       flash[:notice] = "\"#{@wiki.title}\" was deleted successfully."
-      redirect_to [@wiki]
+      redirect_to wikis_path
     else
       flash[:error] = "There was an error deleting the wiki."
       render :show
@@ -60,15 +65,6 @@ class WikisController < ApplicationController
 
   def wiki_params
     params.require(:wiki).permit(:title, :body, :private)
-  end
-
-  def authorize_user
-    wiki = Wiki.find(params[:id])
-
-    unless current_user == wiki.user || not(wiki.private?)
-      flash[:alert] = "You can only edit public wikis."
-      redirect_to [wiki]
-    end
   end
 
 end
